--- a/Sources/_StringProcessing/ByteCodeGen.swift
+++ b/Sources/_StringProcessing/ByteCodeGen.swift
@@ -107,7 +107,6 @@ fileprivate extension Compiler.ByteCodeGen {
     builder.buildMatchBuiltin(
       cc,
       cc.isStrict(options: options),
-      cc.asciiBitset,
       isScalar: options.semanticLevel == .unicodeScalar)
   }