--- a/Sources/_StringProcessing/Engine/InstPayload.swift
+++ b/Sources/_StringProcessing/Engine/InstPayload.swift
@@ -224,20 +224,11 @@ extension Instruction.Payload {
     return (isScalar: pair.0 == 1, pair.1)
   }
 
-  init(_ cc: BuiltinCC, _ isStrict: Bool, _ isScalar: Bool) {
-    let strictBit = isStrict ? 1 << 15 : 0
-    let scalarBit = isScalar ? 1 << 14 : 0
-    // val must be 16 bits, reserve the top 2 bits for if it is strict ascii or scalar
-    assert(cc.rawValue <= 0x3F_FF)
-    let val = cc.rawValue + UInt64(strictBit) + UInt64(scalarBit)
-    self.init(val)
+  init(_ cc: _CharacterClassModel.Representation, _ isInverted: Bool, _ isStrict: Bool, _ isScalar: Bool) {
+    self.init(CharacterClassPayload(cc, isInverted, isStrict, isScalar).rawValue)
   }
-  var builtinCCPayload: (cc: BuiltinCC, isStrict: Bool, isScalar: Bool) {
-    let val = self.rawValue
-    let cc = BuiltinCC(rawValue: val & 0x3F_FF)!
-    let isStrict = (val >> 15) & 1 == 1
-    let isScalar = (val >> 14) & 1 == 1
-    return (cc, isStrict, isScalar)
+  var characterClassPayload: CharacterClassPayload{
+    return CharacterClassPayload(rawValue: rawValue & _payloadMask)
   }
   
   init(consumer: ConsumeFunctionRegister) {
@@ -355,3 +346,37 @@ extension Instruction.Payload {
   }
 }
 
+struct CharacterClassPayload: RawRepresentable {
+  let rawValue: UInt64
+  // Layout:
+  // Top three bits are isInverted, isStrict, isScalar
+  // Lower 16 bits are _CCM.Representation
+  static let invertedShift: UInt64 = 55
+  static let strictShift: UInt64 = 54
+  static let scalarShift: UInt64 = 53
+  static let ccMask: UInt64 = 0xFF
+  init(rawValue: UInt64) {
+    assert(rawValue & _opcodeMask == 0)
+    self.rawValue = rawValue
+  }
+  init(_ cc: _CharacterClassModel.Representation, _ isInverted: Bool, _ isStrict: Bool, _ isScalar: Bool) {
+    let invertedBit = isInverted ? 1 << CharacterClassPayload.invertedShift : 0
+    let strictBit = isStrict ? 1 << CharacterClassPayload.strictShift : 0
+    let scalarBit = isScalar ? 1 << CharacterClassPayload.scalarShift : 0
+    assert(cc.rawValue <= CharacterClassPayload.ccMask) //
+    self.init(rawValue: cc.rawValue + UInt64(invertedBit) + UInt64(strictBit) + UInt64(scalarBit))
+  }
+  
+  var isInverted: Bool {
+    (self.rawValue >> CharacterClassPayload.invertedShift) & 1 == 1
+  }
+  var isStrict: Bool {
+    (self.rawValue >> CharacterClassPayload.strictShift) & 1 == 1
+  }
+  var isScalar: Bool {
+    (self.rawValue >> CharacterClassPayload.scalarShift) & 1 == 1
+  }
+  var cc: _CharacterClassModel.Representation {
+    _CharacterClassModel.Representation.init(rawValue: self.rawValue & CharacterClassPayload.ccMask)!
+  }
+}