--- a/Sources/_StringProcessing/Engine/MEBuiltins.swift
+++ b/Sources/_StringProcessing/Engine/MEBuiltins.swift
@@ -1,9 +1,0 @@
-
-
-extension Processor {
-
-
-  mutating func builtinAssertion() {
-    fatalError("TODO: assertions and anchors")
-  }
-}