… Sources/_StringProcessing/Engine/MEProgram.swift
… Sources/_StringProcessing/Engine/Processor.swift
… Sources/_StringProcessing/Engine/Registers.swift
… Sources/_StringProcessing/MatchingOptions.swift
… Sources/_StringProcessing/PrintAsPattern.swift
… Sources/_StringProcessing/Regex/ASTConversion.swift
… Sources/_StringProcessing/Regex/DSLTree.swift
… Sources/_StringProcessing/Unicode/WordBreaking.swift
… Sources/_StringProcessing/Utility/AsciiBitset.swift
… Sources/_StringProcessing/Utility/RegexFactory.swift
… Sources/_StringProcessing/Utility/TypedInt.swift
… Sources/_StringProcessing/_CharacterClassModel.swift
… Tests/RegexTests/CompileTests.swift
… Tests/RegexTests/MatchTests.swift