- Adds `matchBuiltin` to match builtins instead of a consumeFn calling `_CharacterClassModel.matches` in most cases
- Removes `AssertionFunction` and changes `assertBy` to match in `Processor` instead

There are some gains (10% in microbenchmarks on builtin character classes) but the main purpose is to set up for future work

Future work:
- The layout of the payloads for `matchBuiltin` and `assertBy` will be optimized (hopefully soon) along with the rest of the instruction layout. Changes made in this branch and the scalar optimization branch showed that this is a strong next step
- I wasn't able to fully nuke `_CharacterClassModel.matches` but I think we're getting very close to being able to rip out most of it and the consumer interface for `CustomCharacterClass`