@rctcwyvrn 2022-07-06
WIP because I need to figure out these weird performance behaviours

Weird thing number 1: The first commit in this branch (copying over the AsciiBitset changes from the matchScalar PR) is resulting in significant speedups

```
=== Improvements =====================================================================
- emailLookaheadNoMatchesAll              58ms	63ms	-5.05ms		-8.0%
- emailLookaheadNoMatchesFirst            58ms	63ms	-4.95ms		-7.9%
- emailLookaheadAll                       81.9ms	86.8ms	-4.95ms		-5.7%
- basicCCC                                9.63ms	10.7ms	-1.08ms		-10.1%
- caseInsensitiveCCC                      11.3ms	12.3ms	-1.05ms		-8.6%
- invertedCCC                             28.5ms	29.4ms	-946µs		-3.2%
- emailRFCAll                             50.2ms	51.1ms	-942µs		-1.8%
- basicRangeCCC                           10.6ms	11.4ms	-782µs		-6.9%
```

I sanity checked this by trying to minimize the change (not copying it over to a new file, not adding the new scalar api) and got the same result. 

```
diff --git a/Sources/_StringProcessing/ConsumerInterface.swift b/Sources/_StringProcessing/ConsumerInterface.swift
index af46b53..9c41002 100644
--- a/Sources/_StringProcessing/ConsumerInterface.swift
+++ b/Sources/_StringProcessing/ConsumerInterface.swift
@@ -11,6 +11,13 @@
 
 @_implementationOnly import _RegexParser
 
+extension Character {
+   var _singleScalarAsciiValue: UInt8? {
+     guard self != "\r\n" else { return nil }
+     return asciiValue
+   }
+ }
+
```
```
before
      internal func matches(char: Character) -> Bool {
        let ret: Bool
        if let val = char.asciiValue {
          if val < 64 {
            ret = (a >> val) & 1 == 1
          } else {
            ret =  (b >> (val - 64)) & 1 == 1
          }
        } else {
          ret = false
        }

        if isInverted {
          return !ret
        }

        return ret
      }
after
      private func matches(_ val: UInt8) -> Bool {
        if val < 64 {
          return (a >> val) & 1 == 1
        } else {
          return (b >> (val - 64)) & 1 == 1
        }
      }

       internal func matches(char: Character) -> Bool {
         let matched: Bool
         if let val = char._singleScalarAsciiValue {
           matched = matches(val)
         } else {
           matched = false
         }

         if isInverted {
           return !matched
         }
         return matched
       }
```

Some quick profiling seems to indicate that the performance increase came the call to `_singleScalarAsciiValue` being faster than the call to `asciiValue`, even though `_singleScalarAsciiValue` also uses `asciiValue`. My best guess is that it has to do with `Character.asciiValue.getter` being inlined before the change but not after, though I honestly don't know why that would make a difference