@rctcwyvrn 2022-07-06
Weird performance thing number 2: The optimization is causing regressions in regexes that do not produce the new instruction and very little improvement in the ones that do

Comparing with main with the bitset changes, we get major regressions in all of the email regexes and smaller ones in almost every other result
```
Comparing against benchmark result file sanity_check_after_bitset.json
=== Regressions ======================================================================
- htmlFirst                               26.9µs	26.7µs	209ns		0.8%
- emailLookaheadFirst                     18.5µs	18.2µs	250ns		1.4%
- emailRFCFirst                           8.38µs	7.92µs	458ns		5.8%
- cssFirst                                93µs	91.9µs	1.17µs		1.3%
- HangulSyllableFirst                     4.06ms	4.01ms	47.8µs		1.2%
- cssAll                                  4.87ms	4.82ms	47.9µs		1.0%
- caseInsensitiveCCC                      11.5ms	11.4ms	107µs		0.9%
- ReluctantQuantWithTerminalWhole         10.6ms	10.4ms	113µs		1.1%
- basicRangeCCC                           10.7ms	10.5ms	126µs		1.2%
- HangulSyllableAll                       8.41ms	8.24ms	167µs		2.0%
- htmlAll                                 13.5ms	13.3ms	205µs		1.5%
- subtractionCCC                          15.3ms	15.1ms	278µs		1.8%
- intersectionCCC                         15.8ms	15.5ms	278µs		1.8%
- notFoundAll                             15.1ms	14.8ms	285µs		1.9%
- notFoundFirst                           15.1ms	14.8ms	298µs		2.0%
- symDiffCCC                              39.2ms	38.3ms	885µs		2.3%
- invertedCCC                             29.5ms	28.6ms	937µs		3.3%
- emailLookaheadNoMatchesFirst            59.5ms	58.3ms	1.12ms		1.9%
- emailLookaheadNoMatchesAll              59.3ms	58.2ms	1.13ms		1.9%
- emailRFCAll                             51.5ms	50ms	1.53ms		3.1%
- emailLookaheadAll                       83.5ms	81.8ms	1.72ms		2.1%
- emailRFCNoMatchesAll                    110ms	107ms	3.03ms		2.8%
- emailRFCNoMatchesFirst                  110ms	107ms	3.06ms		2.8%
=== Improvements =====================================================================
- GraphemeBreakNoCapAll                   9.2ms	9.45ms	-257µs		-2.7%
- EagarQuantWithTerminalWhole             6.67ms	6.76ms	-83.9µs		-1.2%
- GraphemeBreakNoCapFirst                 79.2µs	82.8µs	-3.54µs		-4.3%
```

- Checking the compilation does confirm that neither email regex is generating the new instruction
- The current benchmark suite doesn't have many benchmarks that mostly use builtin character classes, most mostly use custom classes and only use the builtins for whitespace. Need to add new microbenchmarks and real world benchmarks that use builtins.