@rctcwyvrn 2022-07-06
Removing the remnants of the bitset fast path optimization I was testing seems to have reverted most of the regressions

```
Comparing against benchmark result file sanity_check_after_bitset.json
=== Regressions ======================================================================
- emailRFCFirst                           8.17µs	7.92µs	249ns		3.1%
- invertedCCC                             28.8ms	28.6ms	167µs		0.6%
- emailLookaheadAll                       82.3ms	81.8ms	446µs		0.5%
- emailRFCAll                             50.5ms	50ms	528µs		1.1%
=== Improvements =====================================================================
- GraphemeBreakNoCapAll                   9.02ms	9.45ms	-430µs		-4.6%
- caseInsensitiveCCC                      11ms	11.4ms	-375µs		-3.3%
- basicCCC                                9.54ms	9.9ms	-359µs		-3.6%
- ReluctantQuantWhole                     14ms	14.2ms	-238µs		-1.7%
- EagarQuantWithTerminalWhole             6.54ms	6.76ms	-216µs		-3.2%
- ReluctantQuantWithTerminalWhole         10.3ms	10.4ms	-184µs		-1.8%
- htmlAll                                 13.1ms	13.3ms	-137µs		-1.0%
- intersectionCCC                         15.4ms	15.5ms	-121µs		-0.8%
- basicRangeCCC                           10.5ms	10.5ms	-77.5µs		-0.7%
- cssAll                                  4.74ms	4.82ms	-74µs		-1.5%
- GraphemeBreakNoCapFirst                 78.5µs	82.8µs	-4.25µs		-5.1%
- cssFirst                                90.8µs	91.9µs	-1.13µs		-1.2%
- htmlFirst                               26.4µs	26.7µs	-291ns		-1.1%
```

But why? Why did an extra computed property on `_CharacterClassModel` and an extra computation when emitting a `matchBuiltin` affect all these other benchmarks???? Why are the email regexes still regressing?? (albeit only a little bit now) 

Need to make 100% sure that compile time is not affecting the benchmark times