@rctcwyvrn 2022-07-16
I may have gotten a little carried away while updating this branch to match main (internalization of `_CharacterClassModel` and changes to assertions) and I ended up implementing a new DSLTree node for character classes

It seems to all be working but there's a slight hiccup in inverting character classes that I'll work out on Monday